requirements.txt
datetime
colorama
Pillow
pyinstaller